Fix: GUI fixes

+++ src/bots/move_outcome.rs
@@ -5,10 +5,10 @@ use super::bot_input::BotInput;
 #[derive(Clone, Debug)]
 pub struct MoveOutcome {
     pub move_sequence : Vec<BotInput>,
-    pub lines_cleared : usize, 
-    pub aggregate_height : usize, 
-    pub holes : usize, 
-    pub bumpiness : usize, 
+    pub lines_cleared : f32, 
+    pub aggregate_height : f32, 
+    pub holes : f32, 
+    pub bumpiness : f32, 
     pub is_t_spin : bool, 
 }
